BLUE color scheme & Fixed logo function so the inputted size actually changes the size of the logo
