+++ src/ios/GetGrinnected/GetGrinnected/Assets.xcassets/Colors/ColorBlue.colorset/Contents.json
@@ -5,9 +5,9 @@
         "color-space" : "display-p3",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0.871",
-          "green" : "0.639",
-          "red" : "0.580"
+          "blue" : "0.294",
+          "green" : "0.176",
+          "red" : "0.090"
         }
       },
       "idiom" : "universal"
@@ -23,9 +23,9 @@
         "color-space" : "display-p3",
         "components" : {
           "alpha" : "1.000",
-          "blue" : "0.367",
-          "green" : "0.000",
-          "red" : "0.001"
+          "blue" : "0.294",
+          "green" : "0.176",
+          "red" : "0.090"
         }
       },
       "idiom" : "universal"
